Successfully fix the problem while editing the gesture in the tableview
A simple switch between the lines 63-64 solve the problem. The signal
'selectionChanged' of the gesturesITemSelectionModel doesn't work
because I set the setSelectionModel first of the gestureTableView. So
when you want to implement the signal of the QItemSelectionModel, just
'setModel' first to get the data then 'setSelectionModel' of the
gesturesTableView.

The Model/View concept of Qt is somewhat a daunting paradigm to
comprehend but if you understand what it can do then your making
yourself a valuable developer in the near future. #lessonlearned