Fixed the question mark '?' fiasco
Whew! After reading some open issues in the main keyboard repo, I was
able to make keyboard 'type' the question mark '?' character. It is
currently an issue in version 0.11 of keyboard and the workaround can be
found in its issue #84.

Due to the super readability of Python, I was able to solve this
'problem'. Reading someone else code will somehow make you good
developer. #lessonlearned