Apply validation when updating gestures
Took me a while to implement this because every time I update a record without changing the shorthand it always raises the duplicate gesture business rule. I needed to create a separate validateGestureOnUpdate() function to get the current selected index so that I can check if the shorthands for the current selected gesture and existing gesture are different.

So far the updating of gesture is working as expected but needs refactoring.

Enjoy your weekend everybody!