Use string.Template class in About dialog's content
The versions variable is now getting its value from the library itself and no longer hard-coded.

I think the app is 85% complete. The last feature that I want to add is the status bar for the main window. After that more refactoring and testing before the final release.

Happy weekend for now!